requirements.txt
tensorflow
scikit-learn
numpy
glob2
pandas
matplotlib
pillow
joblib
cuda-python